Add bunch of clang warnings

+++ queso/solvers/nnls.cpp
@@ -35,9 +35,9 @@ double NNLS::solve(MatrixType& A, VectorType& B, VectorType& X) {
     double *X_double = X.data();
 
     // Construct buffers
-    double *W = new double[n];
-    double *ZZ = new double[m];
-    int *index = new int[n];
+    double *W = new double[static_cast<std::size_t>(n)];
+    double *ZZ = new double[static_cast<std::size_t>(m)];
+    int *index = new int[static_cast<std::size_t>(n)];
 
     double Rnorm = 0; // Residual norm
